feat: 키워드 개수 제한 및 컴포넌트 분리
